refactored EventForm to its own reusable component
