+++ src/components/EventScreens/CreateEventScreen.js
@@ -1,50 +1,10 @@
 import React from 'react'
 import { connect } from 'react-redux'
 import { createEvent } from '../../reducers/eventReducer'
-import t from 'tcomb-form-native'
-import { ScrollView, View } from 'react-native'
-import { Button } from 'react-native-elements'
-import { paddingSides } from '../../styles/global'
-import { now, inOneHour, formatDate } from '../../utils/dates'
-
-const { Form } = t.form
-
-const options = {
-  fields: {
-    description: {
-      label: 'Kuvaus',
-      error: 'Kuvaus ei saa olla tyhjä.'
-    },
-    startdate: {
-      label: 'Aloitusaika',
-      mode: 'datetime',
-      config: {
-        format: (date) => formatDate(date)
-      }
-    },
-    enddate: {
-      label: 'Lopetusaika',
-      mode: 'datetime',
-      error: 'Tapahtuma ei voi loppua ennen alkamisaikaa.',
-      config: {
-        format: (date) => formatDate(date)
-      }
-    }
-  }
-}
-
-const style = {
-  container: {
-    alignItems: 'center',
-    marginTop: 20
-  },
-  scrollview: {
-    width: '100%',
-    backgroundColor: 'white',
-    paddingLeft: paddingSides,
-    paddingRight: paddingSides
-  }
-}
+import { View, ScrollView } from 'react-native'
+import { now, inOneHour } from '../../utils/dates'
+import EventForm from '../simple/EventForm'
+import styles from '../../styles/global'
 
 class CreateEventScreen extends React.Component {
   constructor(props) {
@@ -73,31 +33,18 @@ class CreateEventScreen extends React.Component {
   render() {
     const { event } = this.state
 
-    const DateIsAfter = t.refinement(t.Date, (date) => date >= event.startdate)
-
-    const Event = t.struct({
-      description: t.String,
-      startdate: t.Date,
-      enddate: DateIsAfter
-    })
-
     return (
-      <View style={style.container}>
-        <ScrollView style={style.scrollview}>
-
-          <Form
+      <View style={styles.noPadding}>
+        <ScrollView
+          contentContainerStyle={{ alignItems: 'center' }}
+        >
+          <EventForm
             ref={this.ref}
-            type={Event}
-            options={options}
-            value={event}
-            onChange={(event) => this.setState({ event })}
+            event={event}
+            onFormChange={(event) => this.setState({ event })}
+            onButtonPress={this.createButton}
+            buttonTitle="Luo tapahtuma"
           />
-
-          <Button
-            onPress={this.createButton}
-            title="Luo tapahtuma"
-          />
-
         </ScrollView>
       </View>
     )
